iOS 6.1 Compare
iOS 6.1 Compare

+++ Compare_Latest/PrivateFrameworks/HomeSharing/HSCloudArtworkInfoRequest.h
@@ -10,11 +10,11 @@
 {
 }
 
-+ (id)requestWithDatabaseID:(unsigned int)arg1 itemID:(unsigned int)arg2 sessionID:(unsigned int)arg3;
-+ (id)requestWithDatabaseID:(unsigned int)arg1 itemID:(unsigned int)arg2;
++ (id)requestWithDatabaseID:(unsigned int)arg1 itemID:(unsigned long long)arg2 sessionID:(unsigned int)arg3;
++ (id)requestWithDatabaseID:(unsigned int)arg1 itemID:(unsigned long long)arg2;
 - (id)canonicalResponseForResponse:(id)arg1;
-- (id)initWithDatabaseID:(unsigned int)arg1 itemID:(unsigned int)arg2 sessionID:(unsigned int)arg3;
-- (id)initWithDatabaseID:(unsigned int)arg1 itemID:(unsigned int)arg2;
+- (id)initWithDatabaseID:(unsigned int)arg1 itemID:(unsigned long long)arg2 sessionID:(unsigned int)arg3;
+- (id)initWithDatabaseID:(unsigned int)arg1 itemID:(unsigned long long)arg2;
 
 @end
